Remove impl details from BlockRenderLayerMap
Also move it to client package
