+++ library/blocks/block-extensions/src/main/java/org/quiltmc/qsl/block/extensions/impl/client/package-info.java
@@ -1,0 +1,4 @@
+@ApiStatus.Internal
+package org.quiltmc.qsl.block.extensions.impl.client;
+
+import org.jetbrains.annotations.ApiStatus;
